+++ CHANGELOG.md
@@ -1,4 +1,18 @@
+## 🎉 v0.0.2 `2023-11-12`
+### ✅ 测试
+- 搞定 difference differenceBy differenceWith 三个函数，测试100%覆盖率！ ([ea30a2c](https://github.com/kwooshung/lomind/commit/ea30a2c))
+
+### ✨ 特性
+- 新增 difference、differenceBy、differenceWith，已与lodash一致 ([3304199](https://github.com/kwooshung/lomind/commit/3304199))
+- 增加了 Array 和 Lang 两终分类，实现了部分函数，参数和Lodash一致，实现逻辑和方法，进行了优化。 ([80047b2](https://github.com/kwooshung/lomind/commit/80047b2))
+- 做到这里，感觉直接使用lodash较好，工具库的工作量不小，在此基础上扩展较好 ([a324937](https://github.com/kwooshung/lomind/commit/a324937))
+
+---
+
 ## 🎉 v0.0.1 `2023-11-11`
 ### 🍻 初始
 - 立项，初始化 ([b602f55](https://github.com/kwooshung/lomind/commit/b602f55))
+
+### 📦️ 发布
+- 0.0.1 ([efad67a](https://github.com/kwooshung/lomind/commit/efad67a))